+++ sdk/resourcemanager/resourcegraph/armresourcegraph/response_types.go
@@ -8,10 +8,22 @@
 
 package armresourcegraph
 
+// ClientResourceChangeDetailsResponse contains the response from method Client.ResourceChangeDetails.
+type ClientResourceChangeDetailsResponse struct {
+	// A list of change details associated with a resource over a specific time interval.
+	ResourceChangeDataArray []*ResourceChangeData
+}
+
+// ClientResourceChangesResponse contains the response from method Client.ResourceChanges.
+type ClientResourceChangesResponse struct {
+	// A list of changes associated with a resource over a specific time interval.
+	ResourceChangeList
+}
+
 // ClientResourcesHistoryResponse contains the response from method Client.ResourcesHistory.
 type ClientResourcesHistoryResponse struct {
-	// Anything
-	Interface any
+	// Dictionary of <any>
+	Value map[string]any
 }
 
 // ClientResourcesResponse contains the response from method Client.Resources.
@@ -25,4 +37,10 @@ type OperationsClientListResponse struct {
 	// Result of the request to list Resource Graph operations. It contains a list of operations and a URL link to get the next
 	// set of results.
 	OperationListResult
+}
+
+// QueryClientGenerateQueryResponse contains the response from method QueryClient.GenerateQuery.
+type QueryClientGenerateQueryResponse struct {
+	// Query Generation Response
+	QueryGenerationResponse
 }