+++ sdk/resourcemanager/resourcegraph/armresourcegraph/CHANGELOG.md
@@ -1,5 +1,39 @@
 # Release History
 
+## 0.10.0 (2026-05-25)
+### Breaking Changes
+
+- Field `Interface` of struct `ClientResourcesHistoryResponse` has been removed
+
+### Features Added
+
+- New enum type `ChangeCategory` with values `ChangeCategorySystem`, `ChangeCategoryUser`
+- New enum type `ChangeType` with values `ChangeTypeCreate`, `ChangeTypeDelete`, `ChangeTypeUpdate`
+- New enum type `PropertyChangeType` with values `PropertyChangeTypeInsert`, `PropertyChangeTypeRemove`, `PropertyChangeTypeUpdate`
+- New enum type `Role` with values `RoleAssistant`, `RoleSystem`, `RoleUser`
+- New enum type `StatusCategory` with values `StatusCategoryFailed`, `StatusCategorySucceeded`
+- New enum type `Versions` with values `VersionsV20230901Preview`
+- New function `*Client.ResourceChangeDetails(ctx context.Context, parameters ResourceChangeDetailsRequestParameters, options *ClientResourceChangeDetailsOptions) (ClientResourceChangeDetailsResponse, error)`
+- New function `*Client.ResourceChanges(ctx context.Context, parameters ResourceChangesRequestParameters, options *ClientResourceChangesOptions) (ClientResourceChangesResponse, error)`
+- New function `*ClientFactory.NewQueryClient() *QueryClient`
+- New function `NewQueryClient(credential azcore.TokenCredential, options *arm.ClientOptions) (*QueryClient, error)`
+- New function `*QueryClient.GenerateQuery(ctx context.Context, body QueryGenerationRequest, options *QueryClientGenerateQueryOptions) (QueryClientGenerateQueryResponse, error)`
+- New struct `HistoryContext`
+- New struct `QueryGenerationRequest`
+- New struct `QueryGenerationResponse`
+- New struct `ResourceChangeData`
+- New struct `ResourceChangeDataAfterSnapshot`
+- New struct `ResourceChangeDataBeforeSnapshot`
+- New struct `ResourceChangeDetailsRequestParameters`
+- New struct `ResourceChangeList`
+- New struct `ResourceChangesRequestParameters`
+- New struct `ResourceChangesRequestParametersInterval`
+- New struct `ResourcePropertyChange`
+- New struct `ResourceSnapshotData`
+- New struct `StatusResponse`
+- New field `Value` in struct `ClientResourcesHistoryResponse`
+
+
 ## 0.9.0 (2023-11-24)
 ### Features Added